backend/requirements.txt

fastapi==0.68.0
uvicorn==0.15.0
python-dotenv==0.19.0
looker-sdk==24.20.0
pydantic==1.8.2
python-jose==3.3.0
